requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
urllib3==2.3.0
